+++ packages/w3p/src/errors/index.ts
@@ -5,4 +5,3 @@ export const errors = {
 }
 
 export * from './provider'
-export * from './runtime'