+++ packages/w3p/src/types/provider.ts
@@ -53,10 +53,10 @@ export interface ProviderBase {
   address?: string
   isConnected: boolean
 
-  connect: () => Promise<void>
+  connect?: () => Promise<void>
 
   addChain?: (chain: Chain) => Promise<void>
-  switchChain: (chainId: ChainId) => Promise<void>
+  switchChain?: (chainId: ChainId) => Promise<void>
 
   signAndSendTx?: (txRequestBody: TxRequestBody) => Promise<TransactionResponse>
   signMessage?: (message: string) => Promise<string>
@@ -79,8 +79,8 @@ export interface IProvider extends ProviderBase, ProviderSubscriber {
     listeners?: ProviderListeners,
   ) => Promise<this>
 
+  connect: () => Promise<void>
   chainDetails?: Chain
-  setChainDetails?: (chain: Chain) => void
 }
 
 export interface ProviderProxyConstructor {